enclave-tls: Fix memory leak in enclave-tls-server
Each enclave_tls_negotiate() call will assign a value to the ssl
field in the tls handle. This value is generally the data allocated
by SSL_new(). After multiple calls, the previously allocated ssl
will be leaked because it is overwritten by the later value. Use
a new tls handle for each handshake to solve this issue.

Signed-off-by: Tianjia Zhang <tianjia.zhang@linux.alibaba.com>